==> DSIR-Courses/week-06/6.01-lesson-streamlit/solution-code/notebooks/00-make-author-csv.ipynb <==
# Code to make Author CSV file obtained from Streamlit Lesson master repo  
https://git.generalassemb.ly/DSI-Flex/streamlit_lesson/blob/master/solution-code/notebooks/03-generate_model.ipynb
---

import pandas as pd

def preprocess_gutenberg(file):
    '''Accept a filepath and return a list of paragraphs.'''
    with open(file) as f:
        file_lines = f.read().split('\n\n')
    file_lines = [each.replace('\n', ' ') for each in file_lines]
    return file_lines

def make_author_df(file, author):
    '''Accept a filepath and author name and return a dataframe.'''
    lines = preprocess_gutenberg(file)
    dataframe = pd.DataFrame(lines, columns=['text'])
    dataframe['author'] = author
    return dataframe

austen = make_author_df('../../data/austen.txt', 'Jane Austen')
poe = make_author_df('../../data/poe.txt', 'Edgar Allan Poe')

df = pd.concat([austen, poe])
df = df.dropna()
df.head(3)

df.to_csv('../../data/austen_poe.csv', index=False)
