==> DSIR-Courses/week-06/6.01-lesson-streamlit/solution-code/notebooks/02-read_a_pickle.ipynb <==
# Loading the model in a new notebook

Here, we'll import our fitted pipeline and use it to make predictions about brand new text.

import numpy as np
import pickle

Use a `with` statement to import the pickled pipeline and save:

# path difference b/c solution code
with open('../../models/author_pipe.pkl', 'rb') as pickle_in:
    pipe = pickle.load(pickle_in)

Let's take a look at the pipeline. Does it still have the attributes and methods we'd expect it to have? 🤔

pipe.classes_

# help(pipe)

## Can we generate a prediction?

pipe.predict(["It is a truth universally acknowledged, that a single man in possession of a good fortune, must be in want of a wife"])

pipe.predict(['quote the raven nevermore or something like that'])

## Can we generate predictions interactively?

user_text = input()

print(user_text)

user_text = input('Enter some text:')
prediction = pipe.predict([user_text])[0]
print(f'You write like: {prediction}')
